refactor(project): standardize module architecture and implement Payment Provider pattern
Major structural refactoring to improve separation of concerns and maintainability across core modules.

- **Payments Module:**
  - Implemented the Strategy Pattern for payment gateways.
  - Created new package service.Provider to encapsulate all payment logic.
  - Moved PaymentServiceImpl and specific gateway implementations (StripeProvider, WalletProvider, CashProvider, MockPaymentProvider) into service.Provider to centralize payment processing logic.

- **Admin, KYC, & Earnings Modules:**
  - Enforced consistent Domain-Driven Design (DDD) package structure (controller, domain, dto, epository, service) across all modules.
  - Cleaned up service interfaces and implementations for better dependency injection standards.

- **General:**
  - Improved code navigability and prepared the architecture for easier unit testing and future gateway integrations.

+++ backend/src/main/java/com/deliverXY/backend/NewCode/payments/domain/PromoCodeUsage.java
@@ -6,6 +6,8 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -32,13 +34,13 @@ public class PromoCodeUsage {
     private Delivery delivery;
 
     @Column(name = "discount_amount")
-    private Double discountAmount;
+    private BigDecimal discountAmount;
 
     @Column(name = "original_amount")
-    private Double originalAmount;
+    private BigDecimal originalAmount;
 
     @Column(name = "final_amount")
-    private Double finalAmount;
+    private BigDecimal finalAmount;
 
     @Column(name = "used_at")
     private LocalDateTime usedAt;
